added mock data to mockData.js and added describe blocks to test
